Move some general functions to cloud.py
